move oracles back under tokenizer
